fixing bug in API #2
