+++ package.json
@@ -2,9 +2,9 @@
   "name": "be",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
-    "start": "nodemon ./src/app.js",
+    "start": "nodemon ./src/index.js",
     "test": "jest",
     "test:watch": "jest --watch",
     "coverage": "jest --coverage"
@@ -14,6 +14,7 @@
   "dependencies": {
     "@ngrok/ngrok": "^1.4.1",
     "bcrypt": "^5.1.1",
+    "chalk": "^4.1.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.0",
